refactor: KEEP-946 move fixture planner from tests to lib/test-data
planPhaseFixtures has a single type-only import from the protocol
registry, so it can live in lib/test-data where the coverage report CLI
reaches it without importing from tests/. This removes the need for the
tsconfig exclude and .dockerignore entry that existed only because the
script pulled the planner out of the excluded tests/ tree.
run-fixture.ts keeps re-exporting it so existing import sites are
unchanged.
