+++ tsconfig.json
@@ -45,7 +45,6 @@
     "scripts/seed/fixtures/dev-workflows.ts",
     "scripts/dev-mint-session.ts",
     "scripts/dev-login.ts",
-    "scripts/dev-login-browser.ts",
-    "scripts/protocol-coverage-report.ts"
+    "scripts/dev-login-browser.ts"
   ]
 }
